example/docker/flask_redis_pg_nginx_docker-compose/app1/requirements.txt
psycopg2-binary
redis==4.1.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
